# app/app/docs/swagger.yaml
swagger: "2.0"
info:
  description: "API documentation for AE2I application"
  version: "1.0.0"
  title: "AE2I API"
host: "localhost:5000"
basePath: "/api"
schemes:
  - "http"
securityDefinitions:
  BearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Ajouter 'Bearer <votre_token>' dans le champ Authorization pour authentifier les requêtes."
...
paths:
...
  /polls:
    get:
      tags:
        - "polls"
      summary: "Get all polls"
      responses:
        200:
          description: "A list of polls"
        500:
          description: "Internal server error"
    post:
      tags:
        - "polls"
      summary: "Create a new poll"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
              choices:
                type: "array"
                items:
                  type: "object"
                  properties:
                    choice_text:
                      type: "string"
      security:
        - BearerAuth: []
      responses:
        201:
          description: "Poll created successfully"
        500:
          description: "Internal server error"
  /polls/{poll_id}:
    get:
      tags:
        - "polls"
      summary: "Get poll by ID"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Poll found"
        404:
          description: "Poll not found"
        500:
          description: "Internal server error"
    put:
      tags:
        - "polls"
      summary: "Update poll"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Poll updated successfully"
        404:
          description: "Poll not found"
        500:
          description: "Internal server error"
    delete:
      tags:
        - "polls"
      summary: "Delete poll"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Poll deleted successfully"
        404:
          description: "Poll not found"
        500:
          description: "Internal server error"
  /polls/{poll_id}/responses:
    get:
      tags:
        - "polls"
      summary: "Get responses by poll ID"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Responses found"
        500:
          description: "Internal server error"
    post:
      tags:
        - "polls"
      summary: "Create response"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              choice:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        201:
          description: "Response created successfully"
        500:
          description: "Internal server error"
  /auth/register:
    post:
      tags:
        - "auth"
      summary: "Register a new user"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
              firstName:
                type: "string"
              lastName:
                type: "string"
              password:
                type: "string"
      responses:
        201:
          description: "User registered successfully"
        400:
          description: "Validation error"
        500:
          description: "Internal server error"
  /auth/login:
    post:
      tags:
        - "auth"
      summary: "Login a user"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
              password:
                type: "string"
      responses:
        200:
          description: "Login successful"
        400:
          description: "Validation error"
        401:
          description: "Invalid email or password"
        500:
          description: "Internal server error"
...